Resolve all clippy warnings; workspace is now clippy-clean
Pre-existing warnings, fixed in the appropriate way for each:

- 5 'unnecessary_cast' on '<spec>.buffer_len as u64' (authenticated_wrap_ops,
  kem_ops, message_ops x3): buffer_len is already u64 (CkOutputBufferSpec) and
  capped_output_len takes u64, so the casts were genuinely redundant -> removed.

- 2 'unnecessary_cast' on 'CKF_VERIFY/CKF_DIGEST as u64' (crypto_ops): CK_FLAGS is
  CK_ULONG (u64 on 64-bit, u32 on 32-bit), so the cast is a no-op on 64-bit but a
  needed widening on 32-bit. Kept the casts (removing them breaks 32-bit; .into()
  would trip useless_conversion on 64-bit) and scoped #[allow(unnecessary_cast)]
  to the two tiny cancel methods with a portability comment.

- 'result_unit_err' on begin_operation_capped: it is a crate-internal helper, so
  narrowed it to pub(crate); the Err(()) at-capacity signal needs no richer type
  and the lint no longer fires on non-exported API.

- 'collapsible_if' in trace_id: collapsed the nested if into a let-chain (MSRV
  1.88 already uses let-chains elsewhere).

- 'err_expect' in a mechanism_registry_source test: .err().expect() -> .expect_err().

- 'await_holding_lock' x2 in the latency bench: the std Mutex around the client
  was unnecessary in a single-threaded sequential block_on loop; own the client
  directly instead of locking across .await.

cargo clippy --workspace --all-targets is clean. Build + tests unchanged
(backend 236, server lib 291).

+++ crates/backend/src/ffi/kem_ops.rs
@@ -62,7 +62,7 @@ impl FfiBackend {
             }
         } else {
             // Data query: allocate caller-specified buffer
-            let capped = super::call_helpers::capped_output_len(spec.buffer_len as u64);
+            let capped = super::call_helpers::capped_output_len(spec.buffer_len);
             out_len = capped as cryptoki_sys::CK_ULONG;
             let mut buf = vec![0u8; capped];
             let rv = unsafe {
