+++ crates/backend/src/ffi/message_ops.rs
@@ -928,7 +928,7 @@ impl FfiBackend {
                 Err(CkRv(rv))
             }
         } else {
-            let capped = super::call_helpers::capped_output_len(output_spec.buffer_len as u64);
+            let capped = super::call_helpers::capped_output_len(output_spec.buffer_len);
             out_len = capped as cryptoki_sys::CK_ULONG;
             let mut buf = vec![0u8; capped];
             let rv = call(&mut ck_params, buf.as_mut_ptr(), &mut out_len);
@@ -1041,7 +1041,7 @@ impl FfiBackend {
                 Err(CkRv(rv))
             }
         } else {
-            let capped = super::call_helpers::capped_output_len(output_spec.buffer_len as u64);
+            let capped = super::call_helpers::capped_output_len(output_spec.buffer_len);
             out_len = capped as cryptoki_sys::CK_ULONG;
             let mut buf = vec![0u8; capped];
             let rv = call(&mut ck_params, buf.as_mut_ptr(), &mut out_len);
@@ -1133,7 +1133,7 @@ impl FfiBackend {
                 Err(CkRv(rv))
             }
         } else {
-            let capped = super::call_helpers::capped_output_len(output_spec.buffer_len as u64);
+            let capped = super::call_helpers::capped_output_len(output_spec.buffer_len);
             out_len = capped as cryptoki_sys::CK_ULONG;
             let mut buf = vec![0u8; capped];
             let rv = call(&mut ck_params, buf.as_mut_ptr(), &mut out_len);
